Changed the number format from double to float

+++ exaltPawarikanda/src/main/java/za/co/discovery/assignment/domain/Edge.java
@@ -23,5 +23,5 @@ public class Edge {
     private int id;
     private String startNode;
     private String endNode;
-    private Double distance;
+    private Float distance;
 }
